Move screens into feature packages
