+++ mobile/src/main/java/com/scrolless/app/feature/settings/SettingsScreen.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.scrolless.app.ui.settings
+package com.scrolless.app.feature.settings
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
